output: flush complex nested values, not just map[string]string.
send the full nested complex values when flushing to output plugins
in `Message.Record` as a `map[string]interface{}` instead of as a
simple `map[string]string`. This will allow receiving float64 and
other simple types as well as nested records with key/value pairs
and arrays.

tests have also been added for this new functionality.

Signed-off-by: Phillip Whelan <phil@calyptia.com>