cshared: Make longer liveness of custom plugin
This is because this context cancellation causes immediate cancelling
for Init function of ctx.Done().
To clean processing of custom plugin should be handled by ctx.Done().
We can introduce another way of monitoring the liveness of custom plugin
of Init function of gouroutines.
But it tends to introduce bugs. So, we should change the behavior of
context cancel for custom plugin.

Signed-off-by: Hiroshi Hatake <hiroshi@chronosphere.io>